pyproject.toml
[project]
name = "formula-1-database-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.2",
]
